+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v1.2.1 - [Unreleased]
 ----
  * Added `CollectionComparisonOperator` as default comparison operator for array and enums
  * Added support to use custom namespaces for queries and mutations
+ * Allow set custom type name and description for mutations `input` argument
  * minors bugs fixed
   
 v1.2.0 - 2018-07-09